requirements-dev.txt
# Development dependencies
-r requirements.txt

# Code formatting and linting
black==23.11.0
isort==5.12.0
flake8==6.1.0
flake8-bugbear==23.7.10
flake8-comprehensions==3.12.0
mypy==1.5.1
types-requests==2.31.0.1
types-python-dateutil==2.8.19.14

# Testing
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0

# Pre-commit
pre-commit==3.5.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.1
mkdocstrings[python]==0.23.0

# Development tools
ipython==8.17.2
python-dotenv==1.0.0

# Type checking
mypy-extensions==1.0.0
typing-extensions==4.8.0

# Code quality
bandit==1.7.5
safety==2.3.5

# API documentation
fastapi-chameleon==0.3.5
fastapi-versioning==0.10.0
